payment/requirements.txt
asgiref==3.4.1
Django==3.2.7
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.2
djangorestframework==3.12.4
requests==2.25.1